Create GameBoard
I tried to make a class for the gameboard.  its not done yet but before continuing i wanted to make sure that the board would be drawn to the screen. For some reason i keep getting an array out of bounds exception at line 313.  Any suggestions? also if anyone has any better ideas for how to organize the gameboard instead of using a multidimential array, let me know